restructured the files, puting country in a folder and otp in a folder
