+++ src/screens/signup/phoneVerification/verification/display/index.tsx
