+++ src/screens/signup/phoneVerification/verification/display/style.ts
